Remove Algorand API token
The API token was revoked right after it was introduced in the code, and in fact it was never actually used. But the value remained, which made some people think they found a security vulnerabilities in Ledger Live. Remove the API token to prevent more false-positive security reports from happening.